# docs/api/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_anychat_server_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@anychat.example.com
    name: AnyChat API Support
    url: https://github.com/yzhgit/anychat-server
  description: |-
    AnyChat instant messaging system gateway API service, providing HTTP interfaces for user authentication, user management, and other functions.
    All endpoints requiring authentication must include Authorization: Bearer <token> in the Header.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: AnyChat Gateway API
  version: "1.0"
paths:
...
  /friends:
    get:
      consumes:
      - application/json
      description: Get all friends of current user, supports incremental sync
      parameters:
      - description: last update timestamp (incremental sync)
        format: int64
        in: query
        name: lastUpdateTime
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: get friend list
      tags:
      - friend
  /friends/{id}:
    delete:
      consumes:
      - application/json
      description: Delete specified friend
      parameters:
      - description: friend user ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: delete friend
      tags:
      - friend
  /friends/{id}/remark:
    put:
      consumes:
      - application/json
      description: Update remark name for specified friend
      parameters:
      - description: friend user ID
        in: path
        name: id
        required: true
        type: string
      - description: remark info
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: update friend remark
      tags:
      - friend
  /friends/blacklist:
    get:
      consumes:
      - application/json
      description: Get current user's blacklist
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: get blacklist
      tags:
      - friend
    post:
      consumes:
      - application/json
      description: Add specified user to blacklist
      parameters:
      - description: user ID
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: add to blacklist
      tags:
      - friend
  /friends/blacklist/{id}:
    delete:
      consumes:
      - application/json
      description: Remove specified user from blacklist
      parameters:
      - description: blocked user ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: remove from blacklist
      tags:
      - friend
  /friends/requests:
    get:
      consumes:
      - application/json
      description: Get friend request list by numeric request_type
      parameters:
      - default: 1
        description: 'request type: 1-received 2-sent'
        in: query
        name: request_type
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: get friend request list
      tags:
      - friend
    post:
      consumes:
      - application/json
      description: Send friend request to specified user
      parameters:
      - description: request info
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: send friend request
      tags:
      - friend
  /friends/requests/{id}:
    put:
      consumes:
      - application/json
      description: Handle friend request with numeric action
      parameters:
      - description: request ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'handle action: 1-accept 2-reject'
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: handle friend request
      tags:
      - friend
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"